Удали дублирующуюся статическую разметку
