--- js/layouts/question3.js
@@ -1,28 +1,0 @@
-const getOption = (option) => {
-  return `<div class="game__option">
-    <img src="${option.src}" alt="${option.alt}" width="304" height="455">
-  </div>`;
-};
-
-export const question3 = (questionData) => {
-  return `<div class="game">
-    <p class="game__task">Найдите рисунок среди изображений</p>
-    <form class="game__content  game__content--triple">
-      ${questionData.options.map(getOption).join(``)}
-    </form>
-    <div class="stats">
-      <ul class="stats">
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--correct"></li>
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--unknown"></li>
-      </ul>
-    </div>
-  </div>`;
-};
