+++ js/layouts/questionTriple.js
@@ -1,0 +1,14 @@
+const getOption = (option) => {
+  return `<div class="game__option">
+    <img src="${option.src}" alt="${option.alt}" width="304" height="455">
+  </div>`;
+};
+
+export const question3 = (questionData) => {
+  return {
+    title: `Найдите рисунок среди изображений`,
+    optionsRendered: `<form class="game__content  game__content--triple">
+      ${questionData.options.map(getOption).join(``)}
+    </form>`,
+  };
+};
